Adds the option to include the section nav on list pages
The Components section list page was starting to get long, so I though it would be a good idea to add the option to include the section anchor nav on those pages

You can do that now by setting a sectionNav variable in the front matter. The single.html template will always include the section nav still.